requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
openai
python-dotenv
PyMuPDF
spacy
python-multipart